Added basis for notification functionality
Bug - Throws foreignkey 787 error message when adding in attractions with foreignkey higher than 3.

+++ app/src/main/java/app/labs14/roamly/localStorage/AttractionDao.kt
@@ -23,5 +23,5 @@ interface AttractionDao {
     fun getAllAttractions(): LiveData<List<Attraction>>
 
     @Query("SELECT * FROM attraction_table WHERE itinerary_id = :id")
-    fun getAttractions(id: Long): LiveData<List<Attraction>>
+    fun getAttractions(id: Int): LiveData<List<Attraction>>
 }
